// react/07-routing-nextjs/types/note.ts
export const NOTE_TAGS = ["Todo", "Work", "Personal", "Meeting", "Shopping"] as const;
export type NoteTag = typeof NOTE_TAGS[number];

export interface Note {
  id: string;
  title: string;
  content: string;
  createdAt: string;
  updatedAt: string;
  tag: NoteTag;
}

export interface NewNote {
  title: string;
  content: string;
  tag: NoteTag;
}